skills/cursor-mirror/requirements.txt
# cursor-mirror dependencies
# Python >= 3.14 required (uses match/case, type aliases, deferred annotations)
# Install: python3.14 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

PyYAML>=6.0
pytest>=9.0
